% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rhl30_model.R
\name{get_rhl30_model_coef_df}
\alias{get_rhl30_model_coef_df}
\title{Get the RHL30 Model Coefficients}
\usage{
get_rhl30_model_coef_df()
}
\value{
Data frame of the RHL30 coefficients
}
\description{
Return the RHL30 model coefficients as a data frame
}
