% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rhl30_model.R
\name{get_rhl30_scores_df}
\alias{get_rhl30_scores_df}
\title{Generate the RHL30 Predictor Scores}
\usage{
get_rhl30_scores_df(exprs_mat, rhl30_model_df)
}
\arguments{
\item{exprs_mat}{Normalized expression matrix generated by 
`normalize_exprs_mat`}

\item{rhl30_model_df}{Data frame containg details of the RHL30 model. This
should be the data frame from the `get_rhl30_model_coef_df` function.}
}
\value{
A two column data.frame with sample_id and RHL30 scores
}
\description{
`get_rhl30_scores_df` returns RHL30 scores.
}
