% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_exprs_mat.R
\name{get_sample_normalizer_value}
\alias{get_sample_normalizer_value}
\title{Generate the Normalizer Values for each Sample}
\usage{
get_sample_normalizer_value(exprs_mat, normalizer_genes)
}
\arguments{
\item{exprs_mat}{The expression matrix.}

\item{normalizer_genes}{Character vector containing the genes to use for 
calculating the normalizer}
}
\value{
A numeric vector with the samples as names and the normalizer 
  as values
}
\description{
`get_sample_normalizer_value` returns the normalizer values for each 
sample. The normalizer value is based on the geometric mean of the input
`normalizer_genes`
}
