% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_exprs_mat.R
\name{normalize_exprs_mat}
\alias{normalize_exprs_mat}
\title{Normalize the Expression Matrix}
\usage{
normalize_exprs_mat(exprs_mat, sample_normalizer_values)
}
\arguments{
\item{exprs_mat}{The expression matrix}

\item{sample_normalizer_values}{Sample normalizer values generated by 
the `get_sample_normalizer_value` function}
}
\value{
A normalized expression matrix
}
\description{
This function normalizes the expression matrix.
}
